deploy-requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
gunicorn==21.2.0
Pillow==10.0.1
requests==2.31.0
email-validator==2.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21